Prevent multiple arrow key listeners from building up on the document
This fixes a bug in which navigating the searh results with the arrow 
keys, then leaving the search, then coming back, would cause arrow key 
navigation to jump in weird ways. Here's the reproduction steps:

1. open the parsha-picker
2. focus the search input if necessary
3. press `/` to toggle out the parsha picker
4. return to the parsha picker, search something, then try to navigate 
with arrow keys
5. observe jump

I didn't write a UI test for this because the issue occurs in Safari, 
but not in Chrome, which currently is the only browser supported by 
Cypress. Basically, I couldn't get the issue to present in Cypress, so 
there's nothing to test.