Maintain scroll position on screen resize/rotate
Before this change, rotating the screen on mobile was painful because it 
would scroll you to an unpredictable place. This happened because the 
scroll view maintains the same scroll position, but since the sizes of 
the pages change for smaller screens, the scroll value is now in a 
different place.

With this change, we capture how far down the current page you've 
scrolled, and then replay that when the window changes size.

For performance reasons, we debounce the scroll check for 1 second.