View Model: Persist AliyahLabeller across pages
This is necessary so label the end of an עלייה when its last פסוק crosses a page boundary; we need to remember the previous פסוק so we can label its end.

As an unexpected side-effect, this commit (actually the previous commit, but it only appears in tests in this commit) renders an entire extra page for תענית ציבור, just to render the סוף מפטיר label that ends at the very end of the previous page.